requirements.txt
asgiref==3.2.10
Django==3.0.7
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
importlib-metadata==1.7.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
Pillow==7.2.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
